[RTD-1228] refactor functions in IDPay utilities
